requirements.txt
selenium>=4.10.0
webdriver-manager>=4.0.1
undetected-chromedriver>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=6.0.1
setuptools>=65.0.0
requests>=2.31.0
